+++ src/test/kotlin/com/wutsi/codegen/kotlin/server/ServerConfigCodeGeneratorTest.kt
@@ -40,6 +40,18 @@ internal class ServerConfigCodeGeneratorTest : AbstractMustacheCodeGeneratorTest
         assertContent("/kotlin/server/database/application-prod.yml", "${context.outputDirectory}/src/main/resources/application-prod.yml")
     }
 
+    @Test
+    fun `generate with cache`() {
+        val openAPI = createOpenAPI()
+        val context = createContext()
+        context.addService(Context.SERVICE_CACHE)
+        getCodeGenerator(context).generate(openAPI, context)
+
+        assertContent("/kotlin/server/cache/application.yml", "${context.outputDirectory}/src/main/resources/application.yml")
+        assertContent("/kotlin/server/cache/application-test.yml", "${context.outputDirectory}/src/main/resources/application-test.yml")
+        assertContent("/kotlin/server/cache/application-prod.yml", "${context.outputDirectory}/src/main/resources/application-prod.yml")
+    }
+
     @ParameterizedTest
     @ValueSource(strings = ["application.yml", "application-test.yml", "application-prod.yml"])
     fun `generate - do not overwrite`(name: String) {